test: move LSP11 Init test in separate test suite
