+++ tests/LSP11BasicSocialRecovery/LSP11BasicSocialRecovery.test.ts
@@ -1,12 +1,7 @@
-import { ethers } from "hardhat";
-import { expect } from "chai";
-import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-
 import {
   LSP11BasicSocialRecovery__factory,
   LSP6KeyManager,
   UniversalProfile,
-  LSP11BasicSocialRecoveryInit__factory,
 } from "../../types";
 
 import {
@@ -18,216 +13,66 @@ import {
 
 import {
   setupProfileWithKeyManagerWithURD,
-  deployProxy,
+  grantPermissionViaKeyManagerFixture,
 } from "../utils/fixtures";
 
-import { combinePermissions } from "../utils/helpers";
+describe("LSP11BasicSocialRecovery with constructor", () => {
+  let context: LSP11TestContext;
 
-import { PERMISSIONS, ERC725YDataKeys } from "../../constants";
+  const buildTestContext = async (): Promise<LSP11TestContext> => {
+    const accounts = await getNamedAccounts();
 
-/**
- * Sets `permissions` for the `addressToGrant` on the `universalProfile`
- * via `lsp6KeyManager`
- */
-async function grantPermissionViaKeyManagerFixture(
-  EOA: SignerWithAddress,
-  universalProfile,
-  lsp6KeyManager,
-  addressToGrant,
-  permissions
-) {
-  const rawPermissionArrayLength = await universalProfile.callStatic[
-    "getData(bytes32)"
-  ](ERC725YDataKeys.LSP6["AddressPermissions[]"].length);
+    const [UP, KM] = await setupProfileWithKeyManagerWithURD(accounts.owner);
 
-  let permissionArrayLength = ethers.BigNumber.from(
-    rawPermissionArrayLength
-  ).toNumber();
+    const universalProfile = UP as UniversalProfile;
+    const lsp6KeyManager = KM as LSP6KeyManager;
 
-  const newPermissionArrayLength = permissionArrayLength + 1;
-  const newRawPermissionArrayLength = ethers.utils.hexZeroPad(
-    ethers.utils.hexValue(newPermissionArrayLength),
-    16
-  );
-
-  const payload = universalProfile.interface.encodeFunctionData(
-    "setData(bytes32[],bytes[])",
-    [
-      [
-        ERC725YDataKeys.LSP6["AddressPermissions[]"].length,
-        ERC725YDataKeys.LSP6["AddressPermissions[]"].index +
-          rawPermissionArrayLength.substring(2),
-        ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
-          addressToGrant.substring(2),
-      ],
-      [newRawPermissionArrayLength, addressToGrant, permissions],
-    ]
-  );
-  await lsp6KeyManager.connect(EOA)["execute(bytes)"](payload);
-}
-
-describe("LSP11BasicSocialRecovery contract", () => {
-  describe("When using LSP11 contract with constructor", () => {
-    const buildTestContext = async (): Promise<LSP11TestContext> => {
-      const accounts = await getNamedAccounts();
-
-      const [UP, KM] = await setupProfileWithKeyManagerWithURD(accounts.owner);
-
-      const universalProfile = UP as UniversalProfile;
-      const lsp6KeyManager = KM as LSP6KeyManager;
-
-      const deployParams = {
-        owner: universalProfile,
-        target: universalProfile,
-      };
-
-      const lsp11BasicSocialRecovery =
-        await new LSP11BasicSocialRecovery__factory(accounts.any).deploy(
-          deployParams.owner.address,
-          deployParams.target.address
-        );
-
-      const lsp11Permissions = combinePermissions(
-        PERMISSIONS.ADDCONTROLLER,
-        PERMISSIONS.EDITPERMISSIONS
-      );
-
-      await grantPermissionViaKeyManagerFixture(
-        accounts.owner,
-        universalProfile,
-        lsp6KeyManager,
-        lsp11BasicSocialRecovery.address,
-        lsp11Permissions
-      );
-
-      return {
-        accounts,
-        lsp11BasicSocialRecovery,
-        deployParams,
-        universalProfile,
-        lsp6KeyManager,
-      };
+    const deployParams = {
+      owner: universalProfile,
+      target: universalProfile,
     };
 
-    describe("When deploying the contract", () => {
-      let context: LSP11TestContext;
+    const lsp11BasicSocialRecovery =
+      await new LSP11BasicSocialRecovery__factory(accounts.any).deploy(
+        deployParams.owner.address,
+        deployParams.target.address
+      );
 
-      before(async () => {
-        context = await buildTestContext();
+    await grantPermissionViaKeyManagerFixture(
+      accounts.owner,
+      universalProfile,
+      lsp6KeyManager,
+      lsp11BasicSocialRecovery.address
+    );
+
+    return {
+      accounts,
+      lsp11BasicSocialRecovery,
+      deployParams,
+      universalProfile,
+      lsp6KeyManager,
+    };
+  };
+
+  before(async () => {
+    context = await buildTestContext();
+  });
+
+  describe("When deploying the contract", () => {
+    describe("When initializing the contract", () => {
+      shouldInitializeLikeLSP11(async () => {
+        const { lsp11BasicSocialRecovery, deployParams } = context;
+        return {
+          lsp11BasicSocialRecovery,
+          deployParams,
+          initializeTransaction:
+            context.lsp11BasicSocialRecovery.deployTransaction,
+        };
       });
-
-      describe("When initializing the contract", () => {
-        shouldInitializeLikeLSP11(async () => {
-          const { lsp11BasicSocialRecovery, deployParams } = context;
-          return {
-            lsp11BasicSocialRecovery,
-            deployParams,
-            initializeTransaction:
-              context.lsp11BasicSocialRecovery.deployTransaction,
-          };
-        });
-      });
-    });
-
-    describe("When testing deployed contract", () => {
-      shouldBehaveLikeLSP11(buildTestContext);
     });
   });
 
-  describe("When using LSP11contract with proxy", () => {
-    const buildTestContext = async (): Promise<LSP11TestContext> => {
-      const accounts = await getNamedAccounts();
-
-      const [UP, KM] = await setupProfileWithKeyManagerWithURD(accounts.owner);
-
-      const universalProfile = UP as UniversalProfile;
-      const lsp6KeyManager = KM as LSP6KeyManager;
-
-      const deployParams = {
-        owner: universalProfile,
-        target: universalProfile,
-      };
-
-      const lsp11BasicSocialRecoveryInit =
-        await new LSP11BasicSocialRecoveryInit__factory(
-          accounts.owner
-        ).deploy();
-
-      const lsp11BasicSocialRecoveryProxy = await deployProxy(
-        lsp11BasicSocialRecoveryInit.address,
-        accounts.owner
-      );
-
-      const lsp11BasicSocialRecovery = lsp11BasicSocialRecoveryInit.attach(
-        lsp11BasicSocialRecoveryProxy
-      );
-
-      const lsp11Permissions = combinePermissions(
-        PERMISSIONS.ADDCONTROLLER,
-        PERMISSIONS.EDITPERMISSIONS
-      );
-
-      await grantPermissionViaKeyManagerFixture(
-        accounts.owner,
-        universalProfile,
-        lsp6KeyManager,
-        lsp11BasicSocialRecovery.address,
-        lsp11Permissions
-      );
-
-      return {
-        accounts,
-        lsp11BasicSocialRecovery,
-        deployParams,
-        universalProfile,
-        lsp6KeyManager,
-      };
-    };
-
-    const initializeProxy = async (context: LSP11TestContext) => {
-      return context.lsp11BasicSocialRecovery["initialize(address,address)"](
-        context.deployParams.owner.address,
-        context.deployParams.target.address
-      );
-    };
-
-    describe("When deploying the contract as proxy", () => {
-      let context: LSP11TestContext;
-
-      before(async () => {
-        context = await buildTestContext();
-      });
-
-      describe("When initializing the contract", () => {
-        shouldInitializeLikeLSP11(async () => {
-          const { lsp11BasicSocialRecovery, deployParams } = context;
-          const initializeTransaction = await initializeProxy(context);
-
-          return {
-            lsp11BasicSocialRecovery,
-            deployParams,
-            initializeTransaction,
-          };
-        });
-      });
-
-      describe("When calling initialize more than once", () => {
-        it("should revert", async () => {
-          await expect(initializeProxy(context)).to.be.revertedWith(
-            "Initializable: contract is already initialized"
-          );
-        });
-      });
-    });
-
-    describe("When testing deployed contract", () => {
-      shouldBehaveLikeLSP11(() =>
-        buildTestContext().then(async (context) => {
-          await initializeProxy(context);
-
-          return context;
-        })
-      );
-    });
+  describe("When testing deployed contract", () => {
+    shouldBehaveLikeLSP11(buildTestContext);
   });
 });
